Set timezone and locale to London / en-GB
Copied the en-US_ransack file to en-GB_ransack too, just in case.

The ultimate setting that controls en-GB status is in fact in config/settings.yml which is not part of the repo, nor controlled by chef as far as I can tell.  During the deploy we'll have to manually set it.