Fix date selector for non-US locales
There is a CoffeeScript file that attempts to set the date format to
ISO 8601 format but even though the code is called it has no effect
because the date picker has already been instantiated.

Rather than trying to get the CoffeeScript to work correctly I've
deleted the file and used the standard jQuery UI localisation files.